requirements.txt
PyYAML==6.0
fonttools==4.40.0
